+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### vNEXT
 
+- **Breaking change:** new name `angular-emojify`
+- Added UMD bundle and ES2015 modules
+- Added AoT support
 
 ### v0.1.0